added new configure option --with-pdflib3
Therefore --with-pdflib will only work with PDFlib 4.x from now on, if one wants
to build PHP with PDFlib V3.x he has to use the new option --with-pdflib3.

This makes configuration of PDFlib much easier, as no additional libraries
are needed anymore. This is possible, because PDFlib has built in libraries
for tiff/jpg/zlib since V4.


git-svn-id: http://svn.php.net/repository/pecl/pdf/trunk@59249 c90b9560-bf6c-de11-be94-00142212c4b1